matching-service/requirements.txt
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
prometheus-flask-exporter==0.22.4
requests==2.31.0
pydantic==2.4.2
flask-pydantic==0.11.0
flask-limiter==3.5.0